// apps/wallet-mobile/src/features/ReviewTransaction/common/types.ts
export interface Anchor {
  anchor_data_hash: string
  anchor_url: URL
}
// ...
export type Certificate =
  | {
      StakeRegistration: StakeRegistration
    }
  | {
      StakeDeregistration: StakeDeregistration
    }
  | {
      StakeDelegation: StakeDelegation
    }
  | {
      PoolRegistration: PoolRegistration
    }
  | {
      PoolRetirement: PoolRetirement
    }
  | {
      GenesisKeyDelegation: GenesisKeyDelegation
    }
  | {
      MoveInstantaneousRewardsCert: MoveInstantaneousRewardsCert
    }
  | {
      CommitteeHotAuth: CommitteeHotAuth
    }
  | {
      CommitteeColdResign: CommitteeColdResign
    }
  | {
      DRepDeregistration: DRepDeregistration
    }
  | {
      DRepRegistration: DRepRegistration
    }
  | {
      DRepUpdate: DRepUpdate
    }
  | {
      StakeAndVoteDelegation: StakeAndVoteDelegation
    }
  | {
      StakeRegistrationAndDelegation: StakeRegistrationAndDelegation
    }
  | {
      StakeVoteRegistrationAndDelegation: StakeVoteRegistrationAndDelegation
    }
  | {
      VoteDelegation: VoteDelegation
    }
  | {
      VoteRegistrationAndDelegation: VoteRegistrationAndDelegation
    }
// ...
export type DataOption =
  | {
      DataHash: string
    }
  | {
      Data: string
    }
export type ScriptRef =
  | {
      NativeScript: NativeScript
    }
  | {
      PlutusScript: string
    }
export type Mint = [string, MintAssets][]
export type NetworkId = 'Testnet' | 'Mainnet'
export type TransactionOutputs = TransactionOutput[]
export type CostModel = string[]
export type Voter =
  | {
      ConstitutionalCommitteeHotCred: CredType
    }
  | {
      DRep: CredType
    }
  | {
      StakingPool: string
    }
export type VoteKind = 'No' | 'Yes' | 'Abstain'
export type GovernanceAction =
  | {
      ParameterChangeAction: ParameterChangeAction
    }
  | {
      HardForkInitiationAction: HardForkInitiationAction
    }
  | {
      TreasuryWithdrawalsAction: TreasuryWithdrawalsAction
    }
  | {
      NoConfidenceAction: NoConfidenceAction
    }
  | {
      UpdateCommitteeAction: UpdateCommitteeAction
    }
  | {
      NewConstitutionAction: NewConstitutionAction
    }
  | {
      InfoAction: InfoAction
    }
// ...
export interface TransactionBody {
  auxiliary_data_hash?: string | null
  certs?: Certificate[] | null
  collateral?: TransactionInput[] | null
  collateral_return?: TransactionOutput | null
  current_treasury_value?: string | null
  donation?: string | null
  fee: string
  inputs: TransactionInput[]
  mint?: Mint | null
  network_id?: NetworkId | null
  outputs: TransactionOutputs
  reference_inputs?: TransactionInput[] | null
  required_signers?: string[] | null
  script_data_hash?: string | null
  total_collateral?: string | null
  ttl?: string | null
  update?: Update | null
  validity_start_interval?: string | null
  voting_procedures?: VoterVotes[] | null
  voting_proposals?: VotingProposal[] | null
  withdrawals?: {
    [k: string]: string
  } | null
}
export interface StakeRegistration {
  coin?: string | null
  stake_credential: CredType
}
export interface StakeDeregistration {
  coin?: string | null
  stake_credential: CredType
}
export interface StakeDelegation {
  pool_keyhash: string
  stake_credential: CredType
}
export interface PoolRegistration {
  pool_params: PoolParams
}
// ...
export interface PoolRetirement {
  epoch: number
  pool_keyhash: string
}
export interface GenesisKeyDelegation {
  genesis_delegate_hash: string
  genesishash: string
  vrf_keyhash: string
}
export interface MoveInstantaneousRewardsCert {
  move_instantaneous_reward: MoveInstantaneousReward
}
export interface MoveInstantaneousReward {
  pot: MIRPot
  variant: MIREnum
}
export interface StakeToCoin {
  amount: string
  stake_cred: CredType
}
export interface CommitteeHotAuth {
  committee_cold_credential: CredType
  committee_hot_credential: CredType
}
export interface CommitteeColdResign {
  anchor?: Anchor | null
  committee_cold_credential: CredType
}
export interface DRepDeregistration {
  coin: string
  voting_credential: CredType
}
export interface DRepRegistration {
  anchor?: Anchor | null
  coin: string
  voting_credential: CredType
}
export interface DRepUpdate {
  anchor?: Anchor | null
  voting_credential: CredType
}
export interface StakeAndVoteDelegation {
  drep: DRep
  pool_keyhash: string
  stake_credential: CredType
}
export interface StakeRegistrationAndDelegation {
  coin: string
  pool_keyhash: string
  stake_credential: CredType
}
export interface StakeVoteRegistrationAndDelegation {
  coin: string
  drep: DRep
  pool_keyhash: string
  stake_credential: CredType
}
export interface VoteDelegation {
  drep: DRep
  stake_credential: CredType
}
export interface VoteRegistrationAndDelegation {
  coin: string
  drep: DRep
  stake_credential: CredType
}
export interface TransactionInput {
  index: number
  transaction_id: string
}
export interface TransactionOutput {
  address: string
  amount: Value
  plutus_data?: DataOption | null
  script_ref?: ScriptRef | null
}
export interface Value {
  coin: string
  multiasset?: MultiAsset | null
}
export interface MultiAsset {
  [k: string]: Assets
}
export interface MintAssets {
  [k: string]: string
}
export interface Update {
  epoch: number
  proposed_protocol_parameter_updates: {
    [k: string]: ProtocolParamUpdate
  }
}
export interface ProtocolParamUpdate {
  ada_per_utxo_byte?: string | null
  collateral_percentage?: number | null
  committee_term_limit?: number | null
  cost_models?: Costmdls | null
  d?: UnitInterval | null
  drep_deposit?: string | null
  drep_inactivity_period?: number | null
  drep_voting_thresholds?: DRepVotingThresholds | null
  execution_costs?: ExUnitPrices | null
  expansion_rate?: UnitInterval | null
  extra_entropy?: Nonce | null
  governance_action_deposit?: string | null
  governance_action_validity_period?: number | null
  key_deposit?: string | null
  max_block_body_size?: number | null
  max_block_ex_units?: ExUnits | null
  max_block_header_size?: number | null
  max_collateral_inputs?: number | null
  max_epoch?: number | null
  max_tx_ex_units?: ExUnits | null
  max_tx_size?: number | null
  max_value_size?: number | null
  min_committee_size?: number | null
  min_pool_cost?: string | null
  minfee_a?: string | null
  minfee_b?: string | null
  n_opt?: number | null
  pool_deposit?: string | null
  pool_pledge_influence?: UnitInterval | null
  pool_voting_thresholds?: PoolVotingThresholds | null
  protocol_version?: ProtocolVersion | null
  ref_script_coins_per_byte?: UnitInterval | null
  treasury_growth_rate?: UnitInterval | null
}
// ...
export interface VoterVotes {
  voter: Voter
  votes: Vote[]
}
export interface Vote {
  action_id: GovernanceActionId
  voting_procedure: VotingProcedure
}
export interface GovernanceActionId {
  index: number
  transaction_id: string
}
export interface VotingProcedure {
  anchor?: Anchor | null
  vote: VoteKind
}
export interface VotingProposal {
  anchor: Anchor
  deposit: string
  governance_action: GovernanceAction
  reward_account: string
}